Add files via upload
Used for centering the linear rails on the steel tube while drilling and assembling. For Y2 only, after assembling and squaring the frame and installing the X gantry, loosen the Y2 linear rail screws and use the gantry to align the y2 linear rail to the Y1 and retighten as you slide the gantry along.